Valgrindin'
Put the code through valgrind, found 1 huge memory leak in SERVER
code and several small leaks in Josh's codebase. Hope that the
issue with random server segfaults is resolved with this.

The basis for this hope in that the one time I caught this segfault
by gdb, it happened from fprintf's downward callchain. Hence I
suspected stack overflow.