Major bug removal
The client random crashes bug is likely fixed, as is the server bug.
Those involved seem to have involved reading files before the mate
finished writing them. To remove this, functionality is added to
SERVER and CLIENT - temporary file name generation and file
deployment.

This commit is release material. Henceforth it will be referred as
v.1.1.