Preparations for common TAU error dynamics assault
Following Josh's recomendation I cleaned up most known bugs in this
commit. This included rooting out the recurring scores which
individual TAUs used to make and making some debugging messages for
catching the rare-occuring bug 006 red-handed. A lot of cosmetics
was applied in process.

The transition from identity-learning algorithm to the complete
3-layered user-modelling network turned out to be trivial at this
point, so I'll stick with identity learning for a while.