Bug 009 resolution
The problem with OPTIMIZER::Genome_Get_Random_But_Not turned out
to have to do woth the moment it is called, rather than its
definition. Upon its initial call it looked for a random individual
in unscored population, which is not very productive. This is fixed
by the corresponding change in envs.cpp. The function itself is now
less robust, though - this will be reverted in next commit.