Gamma trick implemented
The idea of constraining the TAU error to a certain value
inversely proportional to the number of controllers is implemented
in this commit. A bit of optimization is added to TAUS - now the
updating function TAUS::storePref only constructs common TAU when
TAU is done bisecting (it used to do that for every preference
indicated).