Investigating the cause of common TAU bump effect.
For opposing users, sometimes there is a short period of time during which
common TAU prevails in terms of the backpropagation error. I think that happens
due to the rapid growth of common TAU for "opposing" runs, which overpowers
the error intoduced by conflicts for a short while. To check this, I'll run a
serie of short simulations with precise user models - 4 hidden neurons,
stricter constraints on convergence, more backpropagation iterations.